add lib func razbr_str
